# Improved GUI performance