--- a/TFT/src/User/API/UI/GUI.c
+++ b/TFT/src/User/API/UI/GUI.c
@@ -108,7 +108,7 @@ void GUI_DrawPoint(uint16_t x, uint16_t y)
 void GUI_FillRect(uint16_t sx, uint16_t sy, uint16_t ex, uint16_t ey)
 {
   uint16_t i=0, j=0;
-  LCD_SetWindow( sx, sy, ex-1, ey-1);
+  LCD_SetWindow(sx, sy, ex-1, ey-1);
   for (i=sx; i<ex; i++)
   {
     for (j=sy; j<ey; j++)
@@ -606,49 +606,89 @@ void GUI_DispOne(int16_t sx, int16_t sy, const CHAR_INFO *pInfo)
 {
   if (pInfo->bytes == 0) return;
 
+  uint8_t w = pInfo->pixelWidth;
+  uint8_t h = pInfo->pixelHeight;
+  uint16_t bitMapSize = (h * w / 8);
+  uint8_t font[bitMapSize];
   uint8_t x = 0;
   uint8_t y = 0;
   uint8_t j = 0;
   uint8_t i = 0;
-  uint16_t bitMapSize = (pInfo->pixelHeight * pInfo->pixelWidth / 8);
-  uint8_t  font[bitMapSize];
+  uint8_t jj = (h + 8 - 1) / 8;
+  uint32_t pixel = 1 << (h - 1);
   uint32_t temp = 0;
 
   W25Qxx_ReadBuffer(font, pInfo->bitMapAddr, bitMapSize);
 
-  for (x = 0; x < pInfo->pixelWidth; x++)
+  // NOTE: the following code was split intentionally for speedup performance despite some more flash usage
+
+  if (guiTextMode == GUI_TEXTMODE_TRANS)
   {
-    for (j = 0; j < (pInfo->pixelHeight + 8 - 1) / 8; j++)
+    for (x = 0; x < w; x++)
     {
-      temp <<= 8;
-      temp |= font[i++];
+      for (j = 0; j < jj; j++)
+      {
+        temp <<= 8;
+        temp |= font[i++];
+      }
+
+      for (y = 0; y < h; y++)
+      {
+        if (temp & pixel)  // draw text pixel
+          GUI_DrawPixel(sx, sy + y, foreGroundColor);
+
+        temp <<= 1;
+      }
+
+      sx++;
+    }
+  }
+  else  // if GUI_TEXTMODE_NORMAL or GUI_TEXTMODE_ON_ICON
+  {
+    uint16_t _buf[h * w];
+    uint16_t *buf;
+    GUI_RECT limit = {0};
+
+    if (guiTextMode == GUI_TEXTMODE_NORMAL)
+      buf = _buf;
+    else  // if GUI_TEXTMODE_ON_ICON, use iconBuffer
+      buf = iconBuffer;
+
+    if (pixel_limit_flag == 1)
+    {
+      if (sx < pixel_limit_rect.x0)
+        limit.x0 = pixel_limit_rect.x0 - sx;
+
+      if (sx + w >= pixel_limit_rect.x1)
+        limit.x1 = sx + w - pixel_limit_rect.x1;
+
+      if (sy < pixel_limit_rect.y0)
+        limit.y0 = pixel_limit_rect.y0 - sy;
+
+      if (sy + h >= pixel_limit_rect.y1)
+        limit.y1 = sy + h - pixel_limit_rect.y1;
     }
 
-    for (y = 0; y < pInfo->pixelHeight; y++)
+    for (x = 0; x < w; x++)
     {
-      if (temp & (1 << (pInfo->pixelHeight - 1)))  // draw text pixel
+      for (j = 0; j < jj; j++)
       {
-        GUI_DrawPixel(sx, sy + y, foreGroundColor);
+        temp <<= 8;
+        temp |= font[i++];
       }
-      else  // draw background pixel
+
+      for (y = 0; y < h; y++)
       {
-        switch (guiTextMode)
-        {
-          case GUI_TEXTMODE_NORMAL:
-            GUI_DrawPixel(sx, sy + y, backGroundColor);
-            break;
+        if (temp & pixel)                             // draw text pixel
+          buf[(y * w) + x] = foreGroundColor;
+        else if (guiTextMode == GUI_TEXTMODE_NORMAL)  // draw background pixel
+          buf[(y * w) + x] = backGroundColor;
 
-          case GUI_TEXTMODE_ON_ICON:
-            GUI_DrawPixel(sx, sy + y, iconBuffer[(pInfo->pixelWidth * y) + x]);
-            break;
-
-          default:
-            break;
-        }
+        temp <<= 1;
       }
-      temp <<= 1;
     }
-    sx++;
+
+    lcd_buffer_display(sx, sy, w, h, buf, &limit);
   }
 }