--- a/TFT/src/User/API/menu.c
+++ b/TFT/src/User/API/menu.c
@@ -942,16 +942,17 @@ void showLiveInfo(uint8_t index, const LIVE_INFO * liveicon, bool redrawIcon)
   {
     if (liveicon->enabled[i] == true)
     {
+      uint8_t textWidth = GUI_StrPixelWidth(liveicon->lines[i].text);
       GUI_POINT loc;
 
       switch (liveicon->lines[i].h_align)
       {
         case CENTER:
-          loc.x = liveicon->lines[i].pos.x - GUI_StrPixelWidth(liveicon->lines[i].text) / 2;
+          loc.x = liveicon->lines[i].pos.x - textWidth / 2;
           break;
 
         case RIGHT:
-          loc.x = liveicon->lines[i].pos.x - GUI_StrPixelWidth(liveicon->lines[i].text);
+          loc.x = liveicon->lines[i].pos.x - textWidth;
           break;
 
         default:
@@ -963,7 +964,7 @@ void showLiveInfo(uint8_t index, const LIVE_INFO * liveicon, bool redrawIcon)
       switch (liveicon->lines[i].v_align)
       {
         case CENTER:
-          loc.y = liveicon->lines[i].pos.y  - BYTE_HEIGHT / 2;
+          loc.y = liveicon->lines[i].pos.y - BYTE_HEIGHT / 2;
           break;
 
         case BOTTOM:
@@ -982,12 +983,13 @@ void showLiveInfo(uint8_t index, const LIVE_INFO * liveicon, bool redrawIcon)
       {
         GUI_SetTextMode(liveicon->lines[i].text_mode);
         GUI_SetBkColor(liveicon->lines[i].bk_color);
-      }
 
-      if (redrawIcon || liveicon->iconIndex == ICON_NULL)
         GUI_DispString(iconPt.x + loc.x, iconPt.y + loc.y, liveicon->lines[i].text);
+      }
       else
+      {
         GUI_DispStringOnIcon(liveicon->iconIndex, iconPt, loc, liveicon->lines[i].text);
+      }
     }
   }