@digant73 2022-04-24
@bigtreetech If function LCD_WR_16BITS_DATA() could be also used just to move to next display address without any write of pixel color on the display, then the optimization could be also extended to text with transparency. I tried to make a custom LCD_WR_16BITS_DATA() but I was not able to force only the display address change without also writing a pixel color on the screen.

The idea was something like that:

if (buf[i] == transPixel)  // if it is a transparency pixel, move to next display address
  LCD_WR_16BITS_NO_DATA();  // move to next display address
else
  LCD_WR_16BITS_NO_DATA(buf[i]);  // write pixel color and move to next display address