@ETE-Design 2022-05-02
@digant73 Have an Sidewinder X1 - Have looked at your Thingiverse... And your firmware work great when I use it... But would like to compile my own... What ENVS have you used to Compile? Also is it Bitmap Folder for TFT28 or TFT35 you use?