@ETE-Design 2022-05-03
> @ETE-Design I use vscode.

Same here... But what ENVS have you set in PlatformIO.ini so it work with MKS TFT28 on an X1?