@digant73 2022-05-03
https://github.com/digant73/BIGTREETECH-TouchScreenFirmware/tree/artillery_tft_fw_1.27.x_patch_8
The env is already set as default