@ETE-Design 2022-05-23
@digant73 That worked on my X1... Funny thing is, in the past I updated my X2 TFT Firmware with yours, see the picture :-) Now it won't update when trying to use your newest version in Thingiverse... It only try to update from the Config.ini nothing else, can you help again?
![IMG_20220523_212039](https://user-images.githubusercontent.com/15981309/169891705-01f63769-c6d7-4062-b79c-6ce0bed0077d.jpg)